requirements.txt
torch
torchvision
wandb
pandas
tqdm
kagglehub
python-chess
pyarrow